+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ActFramework Change Log
 
-**1.8.1**
+**1.8.1** 11/Mar/2018
 * Support request forward #556
 * Cannot apply text result to format: javascript #555
 * Random `NullPointerException` on act-1.6.6 #553